refactor: sort files in related folders
