+++ packages/experience-builder-sdk/src/blocks/preview/CompositionBlock.tsx
@@ -2,16 +2,20 @@ import React, { useMemo } from 'react';
 
 import type { UnresolvedLink } from 'contentful';
 import omit from 'lodash.omit';
-import { EntityStore } from '../core/EntityStore';
+import { EntityStore } from '../../core/preview/EntityStore';
 
-import { CF_STYLE_ATTRIBUTES, CONTENTFUL_CONTAINER_ID, CONTENTFUL_SECTION_ID } from '../constants';
+import {
+  CF_STYLE_ATTRIBUTES,
+  CONTENTFUL_CONTAINER_ID,
+  CONTENTFUL_SECTION_ID,
+} from '../../constants';
 import {
   createDesignComponentRegistration,
   getComponentRegistration,
-} from '../core/componentRegistry';
-import { buildCfStyles } from '../core/stylesUtils';
-import { ResolveDesignValueType } from '../hooks/useBreakpoints';
-import { useStyleTag } from '../hooks/useStyleTag';
+} from '../../core/componentRegistry';
+import { buildCfStyles } from '../../utils/stylesUtils';
+import { ResolveDesignValueType } from '../../hooks/useBreakpoints';
+import { useStyleTag } from '../../hooks/useStyleTag';
 import type {
   Breakpoint,
   CompositionDataSource,
@@ -20,12 +24,12 @@ import type {
   CompositionVariableValueType,
   ExperienceEntry,
   StyleProps,
-} from '../types';
-import { ContentfulContainer } from './ContentfulContainer';
-import { transformContentValue } from './transformers';
-import { resolveDesignComponent } from '../core/preview/designComponentUtils';
-import { DesignComponent } from './DesignComponent';
-import { checkIfDesignComponent } from '../utils';
+} from '../../types';
+import { ContentfulContainer } from '../../components/ContentfulContainer';
+import { transformContentValue } from '../../utils/transformers';
+import { resolveDesignComponent } from '../../core/preview/designComponentUtils';
+import { DesignComponent } from '../../components/DesignComponent';
+import { checkIfDesignComponent } from '../../utils/utils';
 
 type CompositionBlockProps = {
   node: CompositionNode;
