+++ packages/experience-builder-sdk/src/blocks/preview/PreviewDeliveryRoot.test.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import { EntityStore } from '../core/EntityStore';
+import { EntityStore } from '../../core/preview/EntityStore';
 
 import { PreviewDeliveryRoot } from './PreviewDeliveryRoot';
-import { Experience, SchemaVersions } from '../types';
-import { createCompositionEntry } from '../../test/__fixtures__/composition';
-import { assets, entries } from '../../test/__fixtures__/entities';
+import { Experience, SchemaVersions } from '../../types';
+import { createCompositionEntry } from '../../../test/__fixtures__/composition';
+import { assets, entries } from '../../../test/__fixtures__/entities';
 import type { Entry } from 'contentful';
-import { compatibleVersions } from '../constants';
-import { defineComponents, resetComponentRegistry } from '../core/componentRegistry';
+import { compatibleVersions } from '../../constants';
+import { defineComponents, resetComponentRegistry } from '../../core/componentRegistry';
 
 const locale = 'en-US';
 const compositionEntry = createCompositionEntry({
