+++ packages/experience-builder-sdk/src/blocks/preview/PreviewDeliveryRoot.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import { EntityStore } from '../core/EntityStore';
-import { Experience, InternalSDKMode } from '../types';
+import { EntityStore } from '../../core/preview/EntityStore';
+import { Experience, InternalSDKMode } from '../../types';
 import { CompositionBlock } from './CompositionBlock';
-import { compatibleVersions } from '../constants';
-import { useBreakpoints, useEditorModeSwitch } from '../hooks';
+import { compatibleVersions } from '../../constants';
+import { useBreakpoints, useEditorModeSwitch } from '../../hooks';
 
 type DeliveryRootProps = {
   experience: Experience<EntityStore>;
