+++ packages/experience-builder-sdk/src/core/editor/EditorModeEntityStore.ts
@@ -1,6 +1,6 @@
 import { EditorEntityStore, RequestedEntitiesMessage } from '@contentful/visual-sdk';
 import type { Asset, AssetFile, Entry, UnresolvedLink } from 'contentful';
-import { sendMessage } from '../communication/sendMessage';
+import { sendMessage } from '../../communication/sendMessage';
 
 export class EditorModeEntityStore extends EditorEntityStore {
   constructor({ entities, locale }: { entities: Array<Entry | Asset>; locale: string }) {
