+++ packages/experience-builder-sdk/src/hooks/useFetchExperience.ts
@@ -1,6 +1,6 @@
 import type { ContentfulClientApi, Entry } from 'contentful';
 import { useCallback, useState } from 'react';
-import { EntityStore } from '../core/EntityStore';
+import { EntityStore } from '../core/preview/EntityStore';
 import { fetchReferencedEntities, fetchExperienceEntry } from '../core/fetchers';
 import { Experience, ExternalSDKMode } from '../types';
 import { createExperience } from '../core';
